Change normal pror to createContext Split local storage code and app functions to context
